server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = ""
authors = [
    {name = "bdyldo",email = "dyl123455@outlook.com"}
]
readme = "../README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "python-socketio (>=5.13.0,<6.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "black (>=25.1.0,<26.0.0)",
    "torch (>=2.7.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

